refactor(storage): move StorageService to interface and update imports
