+++ apps/backend/files-service/src/files/application/file.service.ts
@@ -7,14 +7,14 @@ import {
 } from '@LucidRF/files-contracts';
 import { Inject, Injectable, Logger } from '@nestjs/common';
 import { v4 as uuidv4 } from 'uuid';
+import { StorageService } from '../../storage/interfaces';
 import { STORAGE_BUCKET_NAME } from '../../storage/storage.constants';
-import { StorageService } from '../../storage/storage.service';
 import { CreateFileRepoDto } from '../domain/dtos';
 import { FileEntity, FolderEntity, PermissionEntity, toFileDto } from '../domain/entities';
 import { AccessLevel, ResourceType } from '../domain/enums';
+import { ResourceNotFoundException } from '../domain/exceptions';
 import { FileRepository, FolderRepository, GroupsService } from '../domain/interfaces';
-import { filterTopLevelSharedItems } from '../domain/logic';
-import { calculateInheritedPermissions } from '../domain/logic/permission.logic';
+import { calculateInheritedPermissions, filterTopLevelSharedItems } from '../domain/logic';
 import { AclService } from './acl.service';
 
 @Injectable()
