+++ apps/backend/files-service/src/storage/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './storage.service.interface';
