+++ apps/backend/files-service/src/storage/interfaces/storage.service.interface.ts
