+++ apps/backend/files-service/src/storage/storage.constants.ts
@@ -1,2 +1,3 @@
 export const MINIO_CLIENT = 'MINIO_CLIENT';
 export const STORAGE_BUCKET_NAME = 'STORAGE_BUCKET_NAME';
+export const DEFAULT_PRESIGNED_URL_EXPIRY = 3600;
